api/requirements.txt
alembic==1.8.1
Flask==2.2.2
flask_migrate==3.1.0
flask_sqlalchemy==2.5.1
PyJWT==2.4.0
SQLAlchemy==1.4.41
Werkzeug==2.2.2
WTForms==3.0.1
email_validator